requirements.txt
Flask==3.1.1
Jinja2==3.1.6
itsdangerous==2.2.0
Werkzeug==3.1.3
click==8.2.1
blinker==1.9.0
#bcrypt
python-dotenv
flask-mail
mysql-connector-python==9.3.0
